% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cleaning.R
\name{beep_alert}
\alias{beep_alert}
\title{A silly function to let you know when your code is done running.}
\usage{
beep_alert(x = 0.8, text = "Just finished running!")
}
\arguments{
\item{x}{time, in seconds}

\item{text}{What the computer will say. Input "" for just beeps.}
}
\description{
All this function does is tell you that your code is done running and beeps. Put this line at the end of a long function so you can know your code is done. Ideally will never be used because our code is so efficient and doesn't require human intervention :)
}
