% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cleaning.R
\name{merge_verbose}
\alias{merge_verbose}
\title{Wrapper for merge that is verbose}
\usage{
merge_verbose(
  x,
  y,
  by,
  by.y = by,
  by.x = by,
  all = F,
  all.x = all,
  all.y = all,
  sort = T,
  suffixes = c(".x", ".y"),
  no.dups = T
)
}
\arguments{
\item{x, y}{data frames, or objects to be coerced to one.}

\item{by, by.x, by.y}{specifications of the columns used for merging. See ‘Details’.}

\item{all}{logical; all = L is shorthand for all.x = L and all.y = L, where L is either TRUE or FALSE.}

\item{all.x}{logical; if TRUE, then extra rows will be added to the output, one for each row in x that has no matching row in y.
These rows will have NAs in those columns that are usually filled with values from y. The default is FALSE, so that only rows with data from both x and y are included in the output.}

\item{all.y}{logical; analogous to all.x.}

\item{sort}{logical. Should the result be sorted on the by columns?}

\item{suffixes}{a character vector of length 2 specifying the suffixes to be used for making unique the names of columns in the result which are not used for merging (appearing in by etc).}
}
\value{
A data frame. The rows are by default lexicographically sorted on the common columns, but for sort = FALSE are in an unspecified order. The columns are the common columns followed by the remaining columns in x and then those in y. If the matching involved row names, an extra character column called Row.names is added at the left, and in all cases the result has ‘automatic’ row names.
}
\description{
Lets you know how many rows were dropped if all=F, and how many rows are all NA's if all=T
}
