% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_anne}
\alias{theme_anne}
\title{Anne's favorite ggplot theme}
\usage{
theme_anne(font = "Avenir", size = 10)
}
\arguments{
\item{font}{The font}

\item{size}{Text size}
}
\description{
A pretty simple theme, based on the tufte theme from ggthemes, but with a transparent background and simpler axes
}
