% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_blank}
\alias{theme_blank}
\title{A truly blank ggplot theme}
\usage{
theme_blank(font = "Avenir", size = 10)
}
\arguments{
\item{font}{The font}

\item{size}{Text size}
}
\description{
A really blank theme - no axes, no ticks, no title, no axis labels, no legend, nothing. Mostly useful for maps.
}
